+++ Engine/src/utils.h
@@ -60,7 +60,7 @@ template<typename Type>
 template<typename Type>
 void axrCallDestructor(Type& item) {
     if constexpr (std::is_array_v<Type>) {
-        for (int i = 0; i < axrGetArrayLength<Type>(); ++i) {
+        for (size_t i = 0; i < axrGetArrayLength<Type>(); ++i) {
             axrCallDestructor(item[i]);
         }
     } else {
